# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Line账号分组管理与进线统计系统API",
        "title": "Line账号管理系统API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取分组列表（支持分页和筛选）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "获取分组列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否激活",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索（激活码或备注）",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新分组（自动生成激活码）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "创建分组",
                "parameters": [
                    {
                        "description": "创建分组请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/:id": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新分组信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "更新分组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新分组请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "删除分组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/:id/generate-subaccount-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员可以为任何分组生成子账户登录Token，普通用户只能为自己管理的分组生成Token，用于在新标签页自动登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "生成子账户Token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/:id/regenerate-code": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定分组重新生成激活码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "重新生成激活码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/batch/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量软删除分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "批量删除分组",
                "parameters": [
                    {
                        "description": "批量删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.BatchDeleteGroupsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.BatchOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/batch/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量更新分组的状态、分类、去重范围等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "批量更新分组",
                "parameters": [
                    {
                        "description": "批量更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.BatchUpdateGroupsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.BatchOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有分组分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "获取分组分类列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.CreateGroupRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "account_limit": {
                    "type": "integer",
                    "example": 10
                },
                "category": {
                    "type": "string",
                    "example": "default"
                },
                "dedup_scope": {
                    "type": "string",
                    "enum": [
                        "current",
                        "global"
                    ],
                    "example": "current"
                },
                "description": {
                    "type": "string",
                    "example": "这是一个测试分组"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "login_password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "remark": {
                    "type": "string",
                    "example": "测试分组"
                },
                "reset_time": {
                    "type": "string",
                    "example": "09:00:00"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "invalid_credentials"
                },
                "message": {
                    "type": "string",
                    "example": "用户名或密码错误"
                }
            }
        },
...
        "schemas.GroupListResponse": {
            "type": "object",
            "properties": {
                "account_limit": {
                    "type": "integer",
                    "example": 10
                },
                "activation_code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "category": {
                    "type": "string",
                    "example": "default"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "dedup_scope": {
                    "type": "string",
                    "example": "current"
                },
                "description": {
                    "type": "string",
                    "example": "这是一个测试分组"
                },
                "duplicate_incoming": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_login_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "line_accounts": {
                    "type": "integer",
                    "example": 4
                },
                "line_business_accounts": {
                    "type": "integer",
                    "example": 1
                },
                "online_accounts": {
                    "type": "integer",
                    "example": 3
                },
                "remark": {
                    "type": "string",
                    "example": "测试分组"
                },
                "reset_time": {
                    "type": "string",
                    "example": "09:00:00"
                },
                "today_duplicate": {
                    "type": "integer",
                    "example": 2
                },
                "today_incoming": {
                    "type": "integer",
                    "example": 10
                },
                "total_accounts": {
                    "description": "统计信息",
                    "type": "integer",
                    "example": 5
                },
                "total_incoming": {
                    "type": "integer",
                    "example": 100
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "description": "用户信息",
                    "type": "string",
                    "example": "user001"
                }
            }
        },
...
        "schemas.UpdateGroupRequest": {
            "type": "object",
            "properties": {
                "account_limit": {
                    "type": "integer",
                    "example": 10
                },
                "category": {
                    "type": "string",
                    "example": "default"
                },
                "dedup_scope": {
                    "type": "string",
                    "enum": [
                        "current",
                        "global"
                    ],
                    "example": "current"
                },
                "description": {
                    "type": "string",
                    "example": "这是一个测试分组"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "login_password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "remark": {
                    "type": "string",
                    "example": "测试分组"
                },
                "reset_time": {
                    "type": "string",
                    "example": "09:00:00"
                }
            }
        },
...
        "utils.PaginationResponse": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/utils.Pagination"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}